feat: 차량 조회  API 개발

See merge request s12-fintech-finance-sub1/S12P21A605!135

+++ src/main/java/com/ssafy/marimo/car/domain/Car.java
@@ -61,7 +61,7 @@ public class Car extends BaseTimeEntity {
     private Float fuelEfficiency;
 
     @Column(nullable = false)
-    private Integer fuelLevel;
+    private Float fuelLevel;
 
     @Column(nullable = true)
     private LocalDateTime lastUpdateDate;
@@ -83,7 +83,7 @@ public class Car extends BaseTimeEntity {
                 FuelType fuelType,
                 Integer totalDistance,
                 Float fuelEfficiency,
-                Integer fuelLevel,
+                Float fuelLevel,
                 OBD2Status obd2Status) {
         this.member = member;
         this.brand = brand;
